requirements.txt
langchain==0.0.339
python-dotenv==1.0.0
streamlit==1.29.0
protobuf~=4.25.1